requirements.txt
pyquery==1.4.3
beautifulsoup4==4.9.3
requests==2.25.0
url-parser==3.0.3
psycopg2-binary==2.8.6
python-dotenv==0.15.0
sqlalchemy==1.3.20
sqlalchemy-utils==0.36.8
# bleach==3.2.1
pytest==6.1.2
pytest_mock==3.3.1
pandas==1.1.4
# trafilatura==0.6.1
spacy==2.3.5
# deeppavlov==0.13.0
tldextract==3.1.0
pyyaml==5.3.1
url-normalize==1.4.3
# grequests==0.6.0
aiohttp_requests==0.1.3
aiodns==2.0.0
cchardet==2.1.7
# yappi==1.3.2
aiohttp-client-cache==0.1.2
flashtext==2.7
url_normalize==1.15.0
w3lib==1.22.0